Abstracted converting matrix to pixel
Added a new file: matrix_to_pixel.hpp. It contains all the logic to modify a pixel based on a draw function. gui will simply identify a draw function and tell the matrix_to_pixel to create the correct pixel based on this function. Then, gui will draw that pixel.